* Added simple SVN class to get svn information while executing scripts * Added meta data writer: Writes meta data XML files for each run when it gets submitted for main processing * Post processing will add its information to the meta data file * Option `--nometadata` disables meta data writing

+++ OfflineSubmitScripts/seasons/2016/libs/runs.py
@@ -50,8 +50,10 @@ def submit_run(dbs4_, g, status, DatasetId, QueueId, ExistingChkSums, dryrun, lo
         GCDFileChkSum = str(FileTools(GCDFileName, logger).md5sum())
         
         lnGCDFile = os.path.join(OutputDir,os.path.basename(GCDFileName))
-        lnCmd = "ln -sf %s %s"%(GCDFileName,lnGCDFile)
-        os.system(lnCmd)
+        
+        if not dryrun:
+            lnCmd = "ln -sf %s %s"%(GCDFileName,lnGCDFile)
+            os.system(lnCmd)
     else:
         GCDFileName = ""
